refactor(web-service): update modal infrastructure
