+++ src/web-service/frontend/src/lib/shopplig-list/AddEntryModal.svelte
@@ -2,9 +2,8 @@
     import Add from "../../assets/svg/Add.svelte";
     import Select from "$lib/forms/Select.svelte";
     import Count from "$lib/forms/Count.svelte";
-    import Checkmark from "../../assets/svg/Checkmark.svelte";
-    import {clickOutside} from "../../assets/helper/clickOutside";
     import {handleErrors} from "../../assets/helper/handleErrors";
+    import Modal from "$lib/general/Modal.svelte";
 
     interface NewEntry {
         id: number,
@@ -17,16 +16,11 @@
         count: number,
     }
 
-    let isOpen = false;
+    let isOpen: boolean = false;
     let entry: NewEntry = { id: 0, count: 0, checked: false };
 
     export let listId: number;
     export let currentEntries: ShoppingListEntry[];
-
-    function handleClickOutside(): void {
-        if (! isOpen) return;
-        isOpen = false;
-    }
 
     // Checks if the selected item is already listed in the current shopping list.
     function findExistingListEntry(): ShoppingListEntry | undefined {
@@ -65,23 +59,9 @@
     </span>
 </button>
 
-<div class:hidden={! isOpen} class="bg-black/80 fixed inset-0 w-screen h-screen"></div>
-
-<section
-    use:clickOutside
-    on:click_outside={handleClickOutside}
-    class:hidden={! isOpen}
-    class="fixed inset-x-4 h-min top-1/2 -translate-y-1/2 bg-white rounded-xl px-4 py-6 sm:left-28 sm:right-8 lg:inset-x-0 lg:max-w-[52rem] lg:mx-auto lg:px-6 lg:px-10 xl:max-w-[62rem]">
+<Modal on:submit={submit} bind:isOpen>
     <h3 class="text-lg font-semibold lg:text-xl">Eintrag hinzufügen:</h3>
 
     <Select bind:entryId={entry.id}/>
     <Count bind:count={entry.count}/>
-
-    <button
-        on:click={submit}
-        disabled="{! entry.id || ! entry.count}"
-        class="bg-green-light mt-8 mx-auto text-white rounded-xl px-5 py-2 flex items-center justify-center gap-x-2 transition-all ease-in-out duration-300 hover:bg-green-dark disabled:bg-gray-light disabled:text-gray-dark">
-        <span class="text-sm lg:text-base">Speichern</span>
-        <Checkmark classes="w-5 h-5 stroke-2"/>
-    </button>
-</section>
+</Modal>
